requirements.txt
keras==2.0.4
numpy==1.12.1
pixivpy==3.3.3
